+++ src/components/crossDevice/MobileFlow.js
@@ -42,6 +42,7 @@ class MobileFlow extends Component {
     )
 
     this.props.actions.setClientSuccess(true)
+    this.props.completeStep(data.docPayload)
   }
 
   onUserAnalyticsEvent = (data) => {