+++ CHANGELOG.md
@@ -16,6 +16,7 @@ This project adheres to the Node [default version scheme](https://docs.npmjs.com
 - Internal: Add support for `.env` file
 - Public: Update Proof of Address flow to present Country Select
 - Public: Determine Proof of Address document options from endpoint
+- Public: Fix error when `mobilePhrases` is supplied but `phrases` are not
 
 ### Fixed